refactor: move user settings to layout folder
